+++ android/app/src/main/java/com/ht/MainActivity.kt
@@ -4,6 +4,7 @@ import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
+import android.os.Bundle
 
 class MainActivity : ReactActivity() {
 
@@ -11,6 +12,10 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
+   override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(null)
+   }
+
   override fun getMainComponentName(): String = "HT"
 
   /**
