+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "HT",
-  "displayName": "HT"
+  "displayName": "Hawaii Tourism"
 }
