+++ package.json
@@ -10,8 +10,12 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "react": "18.2.0",
-    "react-native": "0.74.2"
+    "react-native": "0.74.2",
+    "react-native-safe-area-context": "^4.10.5",
+    "react-native-screens": "^3.32.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
